+++ wallet-database/src/dao/bill.rs
@@ -140,7 +140,7 @@ impl BillDao {
         min_value: Option<f64>,
         start: Option<i64>,
         end: Option<i64>,
-        transfer_type: Vec<i64>,
+        transfer_type: Vec<i32>,
         page: i64,
         page_size: i64,
     ) -> Result<Pagination<BillEntity>, crate::Error>
